+++ checkout/src/main/java/com/mercadopago/sdk/android/checkout/data/remote/response/CardBinResponse.kt
@@ -27,6 +27,6 @@ internal data class InstallmentConfigResponse(
 
 internal data class QuotaResponse(
     val installments: Int?,
-    val installmentAmount: Double?,
-    val totalAmount: Double?,
+    val installmentAmount: Float?,
+    val totalAmount: Float?,
 )
